initial spare part list
